Update test_item_response_theory.R
"Plot the information, which tells me which are in the
x-axis gives me more information in terms of discrimination
power of the items (all items). This is important to show design the items
in a way that they focus more or less on certain parameter
configurations, which in the case of the example is
ability. See plot."